separate common and minestom
