+++ common/src/main/java/net/hollowcube/polar/DimensionTypeWrapper.java
@@ -1,0 +1,6 @@
+package net.hollowcube.polar;
+
+public interface DimensionTypeWrapper {
+    int minY();
+    int maxY();
+}
