+++ minestom/src/main/java/net/minestom/server/instance/palette/PolarPaletteAccessWidener.java
