+++ minestom/src/test/java/net/hollowcube/polar/regression/TestIssue17.java
