+++ settings.gradle.kts
@@ -1,2 +1,15 @@
+pluginManagement {
+    plugins {
+        kotlin("jvm") version "2.1.20"
+    }
+}
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
+}
 rootProject.name = "polar"
 
+include(
+    "common",
+    "minestom",
+    "netty"
+)
